backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# LLM Providers
openai==1.12.0
google-generativeai==0.3.2

# PDF Processing
PyPDF2==3.0.1
azure-ai-formrecognizer==3.3.2
azure-core==1.29.5

# Token counting
tiktoken==0.5.2

# Excel Generation
openpyxl==3.1.2
pandas==2.1.4

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0